pyproject.toml
[project]
name = "semantic-alignment-via-nn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gdown>=5.2.0",
    "hydra-core>=1.3.2",
    "polars>=1.21.0",
    "pyarrow>=19.0.0",
    "python-dotenv>=1.0.1",
    "pytorch-lightning>=2.5.0.post0",
    "scikit-learn>=1.6.1",
    "scipy>=1.15.1",
    "seaborn>=0.13.2",
    "timm>=1.0.14",
    "torch>=2.6.0",
    "wandb>=0.19.6",
]

[tool.ruff]
target-version = "py310"
line-length = 79

[tool.ruff.lint]
extend-select = [
      "C4",
      "SIM",
      "UP",
      "TCH",
]

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
docstring-code-format = true
